% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common_carp.R
\name{common_carp}
\alias{common_carp}
\alias{args_common_carp}
\title{Parameterised common carp population model}
\usage{
common_carp(
  k = 50000,
  reproductive = 3:28,
  system = "main_channel",
  n = c(0, 0, 0),
  ntime = 50,
  start = c(1, 1, 1),
  end = c(1, 1, 1),
  add = TRUE
)

args_common_carp(
  n = c(0, 0, 0),
  ntime = 50,
  start = c(1, 1, 1),
  end = c(1, 1, 1),
  add = TRUE
)
}
\arguments{
\item{k}{carrying capacity, defaults to 50000 but a rule-of-thumb
for setting this value is 2 adult females per linear metre of
a large river (e.g. the Murray River)}

\item{reproductive}{integer vector defining the adult stages in the
common carp population. Defaults to \code{3:28}}

\item{system}{one of \code{"main_channel"}, \code{"river_wetland"},
\code{"ephemeral_wetland"}, \code{"permanent_wetland"},
or \code{"floodplain"}, defining the early life survival.
Defaults to "main_channel" and ignores case}

\item{n}{an integer, vector of integers, or list specifying the
number of young-of-year, 1+, or older fish to add or
remove in any given year. If a single integer is provided,
all age classes are assumed to have the same number of
additions or removals. If a vector is provided, it must
have one value for each age class. If a list is provided,
it must have one element for each age class, with a
value for each time step. Addition versus removal is
controlled with \code{add}. Defaults to
\code{c(0, 0, 0))}, which will specify no additions
or removals}

\item{ntime}{number of time steps used in population
simulation. Defaults to \code{50} and is required
to ensure simulated additions or removals are defined
for every simulated time step}

\item{start}{time step at which additions or removals start
if \code{n} is an integer or vector. Defaults to
\code{c(1, 1, 1)}}

\item{end}{time step at which additions or removals finish
if \code{n} is an integer or vector. Defaults to
\code{c(1, 1, 1)}}

\item{add}{logical indicating whether individuals are
added or removed from the population. Defaults to
\code{TRUE}, which defines additions (e.g., stocking). Can
be specified as a single value, in which case all stages
are set equal, as three values, in which case stages can
differ, or as a matrix with three rows and \code{ntime}
columns, in which case the effects of \code{add} can
change through time}
}
\description{
Use a pre-defined population dynamics object to
  simulate population dynamics of common carp (*Cyrpinus carpio*).
  Model parameters are based on existing data sets and
  published studies on common carp.
}
\details{
The \code{common_carp} population model is a mixed
  age- and stage-structured model with 28 classes and includes negative
  density dependence, environmental and demographic
  stochasticity, and optional associations with hydrological conditions
  in rivers and removal efforts.

  The common carp population template requires
  several additional arguments and allows several optional
  arguments. Arguments are described
  individually above.
}
\examples{
# define a basic model for common carp
carp <- common_carp()

# simulate from this model
sims <- simulate(carp, nsim = 100)

# plot the simulated values
plot(sims)
}
