% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estuary_perch.R
\name{estuary_perch}
\alias{estuary_perch}
\alias{args_estuary_perch}
\title{Parameterised estuary perch population model}
\usage{
estuary_perch(k = 30000, n = 0, ntime = 50, start = 1, end = 1)

args_estuary_perch(n = 0, ntime = 50, start = 1, end = 1)
}
\arguments{
\item{k}{carrying capacity}

\item{n}{an integer or list of integers specifying the
number of individuals removed in any given year.
If a single integer is provided,
all years are assumed to have the same number of
additions or removals. If a list is provided,
it must have one value for each year. Defaults to
\code{0}, which will specify no additions}

\item{ntime}{number of time steps used in population
simulation. Defaults to \code{50} and is required
to ensure simulated additions or removals are defined
for every simulated time step}

\item{start}{time step at which additions or removals start
if \code{n} is an integer. Defaults to
\code{1}}

\item{end}{time step at which additions or removals finish
if \code{n} is an integer. Defaults to
\code{1}}
}
\description{
Use a pre-defined population dynamics object to
  simulate population dynamics of estuary perch
  (*Percalates colonorum*). Model parameters are based
  on existing data sets and published studies on estuary perch.
}
\details{
The \code{estuary_perch} population model is a
  stage-structured model with 2 classes and includes negative
  density dependence, environmental and demographic
  stochasticity, and optional associations with hydrological
  conditions, instream and riparian habitat, and predators.

  The estuary perch population template requires
  several additional arguments and allows several optional
  arguments, described individually above.
}
\examples{
# define a basic model for estuary perch
p <- estuary_perch()

# simulate from this model
sims <- simulate(p, nsim = 100)

# plot the simulated values
plot(sims)
}
