% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/macquarie_perch.R
\name{macquarie_perch}
\alias{macquarie_perch}
\alias{args_macquarie_perch}
\title{Parameterised Macquarie perch population model}
\usage{
macquarie_perch(
  k = 1000,
  reproductive = 3:30,
  system = "lake",
  genetic_factor = 1,
  n = c(0, 0, 0),
  ntime = 50,
  start = c(1, 1, 1),
  end = c(1, 1, 1),
  add = TRUE,
  allee_strength = 1,
  contributing_min = 0.75,
  contributing_max = 1,
  recruit_failure = 0
)

args_macquarie_perch(
  n = c(0, 0, 0),
  ntime = 50,
  start = c(1, 1, 1),
  end = c(1, 1, 1),
  add = TRUE,
  allee_strength = 1,
  contributing_min = 0.75,
  contributing_max = 1,
  recruit_failure = 0
)
}
\arguments{
\item{k}{integer carrying capacity (maximum population size)
for the simulated Macquarie perch population. Defaults to 1000}

\item{reproductive}{integer vector defining the adult stages in the
Macquarie perch population. Defaults to \code{3:30}}

\item{system}{one of \code{"lake"} or \code{"river"}, defining
whether covariate associations are based on lake or river
conditions. Covariate associations are included in simulations
only if covariates are provided to \code{\link[aae.pop]{simulate}}
through the \code{args} argument}

\item{genetic_factor}{positive real value representing proportional
changes in early life (egg, larvae, young-of-year) survival,
included here to allow scenarios of genetic mixing of Macquarie
perch populations (e.g., \insertCite{lutz20}{aae.pop.templates}).
Note that non-default values of \code{genetic_factor} must be
provided when defining the population model with
\code{macquarie_perch} and again when defining arguments with
\code{get_args("macquarie_perch")}}

\item{n}{an integer, vector of integers, or list specifying the
number of young-of-year, 2+, and adult fish to add or
remove in any given year. If a single integer is provided,
all age classes are assumed to have the same number of
additions or removals. If a vector is provided, it must
have one value for each age class. If a list is provided,
it must have one element for each age class, with a
value for each time step. Addition versus removal is
controlled with \code{add}. Defaults to
\code{c(0, 0, 0))}, which will specify no additions
or removals}

\item{ntime}{number of time steps used in population
simulation. Defaults to \code{50} and is required
to ensure simulated additions or removals are defined
for every simulated time step}

\item{start}{time step at which additions or removals start
if \code{n} is an integer or vector. Defaults to
\code{c(1, 1, 1)}}

\item{end}{time step at which additions or removals finish
if \code{n} is an integer or vector. Defaults to
\code{c(1, 1, 1)}}

\item{add}{logical indicating whether individuals are
added or removed from the population. Defaults to
\code{TRUE}, which defines additions (e.g., stocking). Can
be specified as a single value, in which case all stages
are set equal, as three values, in which case stages can
differ, or as a matrix with three rows and \code{ntime}
columns, in which case the effects of \code{add} can
change through time}

\item{allee_strength}{strength of Allee effect. Defaults
to \code{1}. See \insertCite{todd_lint15}{aae.pop.templates}
for details}

\item{contributing_min}{minimum proportion of adult females
contributing to reproduction at any given time step.
Defaults to \code{1.0}}

\item{contributing_max}{maximum proportion of adult females
contributing to reproduction at any given time step.
Defaults to \code{1.0}}

\item{recruit_failure}{probability of complete recruitment
failure at any given time step. Defaults to \code{0}}
}
\description{
Use a pre-defined population dynamics object to
  simulate population dynamics of Macquarie perch
  (*Macquaria australasica*). Model parameters are based
  on existing data sets and published studies on Macquarie
  perch.
}
\details{
The \code{macquarie_perch} population model is an
  age-structured model with 30 age classes and includes negative
  and positive density dependence, environmental and demographic
  stochasticity, and optional associations with hydrological conditions
  in rivers or lakes (controlled with the \code{system} parameter)
  and stocking or angling effects.

  The Macquarie perch population template requires
  several additional arguments and allows several optional
  arguments. Arguments are described
  individually above.

  The full population model is developed and described in
  \insertCite{todd_lint15}{aae.pop.templates}
}
\examples{
# define a basic model for Macquarie perch with
#   carrying capacity = 1000
mp <- macquarie_perch(k = 1000)

# simulate from this model
sims <- simulate(mp, nsim = 100)

# plot the simulated values
plot(sims)
}
\references{
\insertRef{todd_lint15}{aae.pop.templates}
}
