% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/murray_cod.R
\name{murray_cod}
\alias{murray_cod}
\title{Parameterised Murray cod population model}
\usage{
murray_cod(k = 20000)
}
\arguments{
\item{k}{carrying capacity}
}
\description{
Use a pre-defined population dynamics object to
  simulate population dynamics of Murray cod
  (*Maccullochella peelii*). Model parameters are based
  on existing data sets and published studies on Murray
  cod.
}
\details{
The \code{murray_cod} population model is a mixed
  age- and stage-structured model with 25 classes and includes negative
  density dependence, environmental and demographic
  stochasticity, and optional associations with hydrological conditions
  in rivers and stocking or angling effects.
}
\examples{
# define a basic model for Murray cod with
#   carrying capacity = 25000
mc <- murray_cod(k = 25000)

# simulate from this model
sims <- simulate(mc, nsim = 100)

# plot the simulated values
plot(sims)
}
