% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/murray_darling_rainbowfish.R
\name{murray_darling_rainbowfish}
\alias{murray_darling_rainbowfish}
\alias{args_murray_darling_rainbowfish}
\title{Parameterised Murray-Darling rainbowfish population model}
\usage{
murray_darling_rainbowfish(k = 10000, ntime = 50)

args_murray_darling_rainbowfish(ntime = 50)
}
\arguments{
\item{k}{carrying capacity. Defaults to \code{10000}}

\item{ntime}{number of time steps used in population
simulation. Defaults to \code{50}}
}
\description{
Use a pre-defined population dynamics object to
  simulate population dynamics of Murray-Darling rainbowfish
  (*Melanotaenia fluviatilis*). Model parameters are based
  on existing data sets and published studies on Murray-Darling rainbowfish.
}
\details{
The \code{murray_darling_rainbowfish} population model is a
  stage-structured model with 5 classes and includes negative
  density dependence, environmental and demographic
  stochasticity, and optional associations with hydrological
  conditions, instream and riparian habitat, and predators.

  The overarching model structure was based the life history of Murray-
  Darling rainbowfish. The primary model parameters (survival, fecundity,
  density dependence) were estimated from observed counts of rainbowfish
  in 5 length classes (0-30 mm, 30-40 mm, 40-50 mm, 50-70 mm, > 70 mm).
  The effects of hydrology, habitat, and predation were based on
  (minimal) published literature on the species and are subject to
  moderate levels of uncertainty.

  The murray_darling_rainbowfish population template does not
  currently require additional arguments.
}
\examples{
# define a basic model for Murray-Darling rainbowfish
p <- murray_darling_rainbowfish()

# simulate from this model
sims <- simulate(p, nsim = 100)

# plot the simulated values
plot(sims)
}
