% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/platypus.R
\name{platypus}
\alias{platypus}
\alias{args_platypus}
\title{Parameterised platypus population model}
\usage{
platypus(k = 400, n = 0, ntime = 50, start = 1, end = 1)

args_platypus(n = 0, ntime = 50, start = 1, end = 1)
}
\arguments{
\item{k}{carrying capacity. Previously estimated densities
suggest 1-7 adult platypus per km in rivers, with larger values
in systems that more closely resemble reference (or natural)
conditions. Defaults to 400, which assumes a 100 km river
reach at 4 individuals per km. The value of \code{k} is
multiplied by \code{0.65} to give adult female carrying
capacity, assuming a 65:35 female:male sex ratio in
adult platypus}

\item{n}{an integer or list of integers specifying the
number of adult platypus to add in any given year.
If a single integer is provided,
all years are assumed to have the same number of
additions or removals. If a list is provided,
it must have one value for each year. Defaults to
\code{0}, which will specify no additions}

\item{ntime}{number of time steps used in population
simulation. Defaults to \code{50} and is required
to ensure simulated additions or removals are defined
for every simulated time step}

\item{start}{time step at which additions or removals start
if \code{n} is an integer. Defaults to
\code{1}}

\item{end}{time step at which additions or removals finish
if \code{n} is an integer. Defaults to
\code{1}}
}
\description{
Use a pre-defined population dynamics object to
  simulate population dynamics of platypus
  (*Ornithorhynchus anatinus*). Model parameters are based
  on existing data sets and published studies on platypus.
}
\details{
The \code{platypus} population model is a
  stage-structured model with 2 classes and includes negative
  density dependence, environmental and demographic
  stochasticity, and optional associations with hydrological
  conditions, instream and riparian habitat, and predators.

  The platypus population template requires
  several additional arguments and allows several optional
  arguments, described individually above.
}
\examples{
# define a basic model for platypus
p <- platypus()

# simulate from this model
sims <- simulate(p, nsim = 100)

# plot the simulated values
plot(sims)
}
