% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estuary_perch.R
\name{estuary_perch}
\alias{estuary_perch}
\alias{args_estuary_perch}
\title{Parameterised estuary perch population model}
\usage{
estuary_perch(k = 30000, n = 0, ntime = 50, start = 1, end = 1)

args_estuary_perch(
  n = c(0, 0, 0),
  ntime = 50,
  start = c(1, 1, 1),
  end = c(1, 1, 1),
  add = TRUE,
  p_capture = 0
)
}
\arguments{
\item{k}{carrying capacity}

\item{n}{an integer, vector of integers, or list specifying the
number of young-of-year, 2+, and adult fish to add or
remove in any given year. If a single integer is provided,
all age classes are assumed to have the same number of
additions or removals. If a vector is provided, it must
have one value for each age class. If a list is provided,
it must have one element for each age class, with a
value for each time step. Addition versus removal is
controlled with \code{add}. Defaults to
\code{c(0, 0, 0))}, which will specify no additions
or removals}

\item{ntime}{number of time steps used in population
simulation. Defaults to \code{50} and is required
to ensure simulated additions or removals are defined
for every simulated time step}

\item{start}{time step at which additions or removals start
if \code{n} is an integer or vector. Defaults to
\code{c(1, 1, 1)}}

\item{end}{time step at which additions or removals finish
if \code{n} is an integer or vector. Defaults to
\code{c(1, 1, 1)}}

\item{add}{logical indicating whether individuals are
added or removed from the population. Defaults to
\code{TRUE}, which defines additions (e.g., stocking). Can
be specified as a single value, in which case all stages
are set equal, as three values, in which case stages can
differ, or as a matrix with three rows and \code{ntime}
columns, in which case the effects of \code{add} can
change through time}

\item{p_capture}{probability of capture by recreational anglers,
defaults to 0, in which case recreational fishing does not
occur}
}
\description{
Use a pre-defined population dynamics object to
  simulate population dynamics of estuary perch
  (*Percalates colonorum*). Model parameters are based
  on existing data sets and published studies on estuary perch.
}
\details{
The \code{estuary_perch} population model is a
  stage-structured model with 2 classes and includes negative
  density dependence, environmental and demographic
  stochasticity, and optional associations with hydrological
  conditions, instream and riparian habitat, and predators.

  The estuary perch population template requires
  several additional arguments and allows several optional
  arguments, described individually above.
}
\examples{
# define a basic model for estuary perch
p <- estuary_perch()

# simulate from this model
sims <- simulate(p, nsim = 100)

# plot the simulated values
plot(sims)
}
