% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/platypus.R
\name{platypus}
\alias{platypus}
\alias{args_platypus}
\title{Parameterised platypus population model}
\usage{
platypus(k = 400)

args_platypus(...)
}
\arguments{
\item{k}{carrying capacity. Previously estimated densities
suggest 1-7 adult platypus per km in rivers, with larger values
in systems that more closely resemble reference (or natural)
conditions. Defaults to 400, which assumes a 100 km river
reach at 4 individuals per km. The value of \code{k} is
multiplied by \code{0.65} to give adult female carrying
capacity, assuming a 65:35 female:male sex ratio in
adult platypus}

\item{n}{an integer, vector of integers, or list specifying the
number of juvenile and adult platypus to add or
remove in any given year. If a single integer is provided,
all age classes are assumed to have the same number of
additions or removals. If a vector is provided, it must
have one value for each age class. If a list is provided,
it must have one element for each age class, with a
value for each time step. Addition versus removal is
controlled with \code{add}. Defaults to
\code{c(0, 0))}, which will specify no additions
or removals}

\item{ntime}{number of time steps used in population
simulation. Defaults to \code{50} and is required
to ensure simulated additions or removals are defined
for every simulated time step}

\item{start}{time step at which additions or removals start
if \code{n} is an integer or vector. Defaults to
\code{c(1, 1)}}

\item{end}{time step at which additions or removals finish
if \code{n} is an integer or vector. Defaults to
\code{c(1, 1)}}

\item{add}{logical indicating whether individuals are
added or removed from the population. Defaults to
\code{TRUE}, which defines additions (e.g., translocation). Can
be specified as a single value, in which case all stages
are set equal, as three values, in which case stages can
differ, or as a matrix with three rows and \code{ntime}
columns, in which case the effects of \code{add} can
change through time}

\item{p_capture}{probability of capture by recreational anglers,
defaults to 0, in which case recreational fishing does not
occur}

\item{slot}{length slot within which Murray cod can legally
be removed by recreational anglers. Defined in millimetres,
defaults to 550-750 mm}

\item{system}{one of \code{"murray"}, \code{"goulburn"}, or
\code{"campaspe"}, defining flow associations based on those
observed in the Murray River, Goulburn River (currently identical
to observations in the Murray River), or the Campaspe River
(currently based on observations in the Broken River). Defaults
to \code{"murray"}}
}
\description{
Use a pre-defined population dynamics object to
  simulate population dynamics of platypus
  (*Ornithorhynchus anatinus*). Model parameters are based
  on existing data sets and published studies on platypus.
}
\details{
The \code{platypus} population model is a
  stage-structured model with 2 classes and includes negative
  density dependence, environmental and demographic
  stochasticity, and optional associations with hydrological
  conditions, instream and riparian habitat, and predators.

The platypus population template requires
  several additional arguments and allows several optional
  arguments. Arguments can be defined with a call to
  \code{get_args("platypus")} and are described
  individually above.
}
\examples{
# define a basic model for platypus
p <- platypus()

# define required arguments
p_args <- get_args("platypus")

# simulate from this model
sims <- simulate(p, nsim = 100, args = p_args)

# plot the simulated values
plot(sims)
}
